cmake: make sphinx use a python-staging folder instead of site-packages
This means that you no longer need to install UPM in oder to be able to use
`make pydoc`. The target pydoc now depends on target doc which with SWIG will
cause python modules to be built, making parralel builds work as expected. We
wipe the python-staging directory on each make pydoc as it's a custom_target
and will rerun at every time anyways, copy is cheap and in case a module gets
deleted it avoids the doc still getting built

Signed-off-by: Brendan Le Foll <brendan.le.foll@intel.com>