bmp280: Initial implementation
The BMP280 is an absolute barometric pressure sensor especially
designed for mobile applications. The sensor module is housed in an
extremely compact 8-pin metal-lid LGA package with a footprint of only
2.0 × 2.5 mm2 and 0.95 mm package height. Its small dimensions and its
low power consumption of 2.7 μA @1Hz allow the implementation in
battery driven devices such as mobile phones, GPS modules or watches.

As the successor to the widely adopted BMP180, the BMP280 delivers
high performance in all applications that require precise pressure
measurement. The BMP280 operates at lower noise, supports new filter
modes and an SPI interface within a footprint 63% smaller than the
BMP180.

This driver supports both I2C and SPI operation.  However SPI is not
functional on Edison with the Arduino breakout board.

Signed-off-by: Jon Trulson <jtrulson@ics.com>