t8100: Initial implementation
This library adds support for the Amphenol Telaire Ventostat T8100
Ventilation Controller.  It is a BACnet MS/TP device.  It requires the
upm::BACNETMSTP driver and the new upm::BACNETUTIL class.

The Telaire Ventostat T8100 reports Temperature, Humidity and CO2
concentrations.  It supports an optional relay with a settable trigger
point.  The unit this driver was tested under did not support the
optional relay.  The temperature range supported is 0-50C, humidity is
0-100% non-condensing, and CO2 range is 0-2000 PPM for the T8100.

This module was developed using the upm::BACNETMSTP library,
based on libbacnet-stack 0.8.3.  Both libbacnet 0.8.3 and the
upm::BACNETMSTP libraries must be present in order to build this
module.

Additionally, the BACNETMSTP library must include the new BACNETUTIL
class.

It was connected using an RS232->RS485 interface.  You cannot use
the built in MCU TTL UART pins for accessing this device -- you
must use a full Serial RS232->RS485 or USB-RS485 interface
connected via USB.

Signed-off-by: Jon Trulson <jtrulson@ics.com>