src/CMakeLists.txt: "export" lib C and C++ targets to the calling modules
With the switch to using a function rather than a macro, $libname is
no longer available to module CMakeLists.txt files.  These are useful
for adding dependencies (for example a module depending on another
module).

Now ${libnamec} (for C library) and ${libnamecxx} (for C++ library)
targets are made available to the module CMakeLists.txt files for
their use.

Signed-off-by: Jon Trulson <jtrulson@ics.com>