upm: enable Elementz IR Proximity sensor library and example
Elementz IR Infrared Proximity /Obstacle Detector Sensor Module

This sensor module works on INFRARED, for obstacles with reflective surfaces
(white colored), the maximum range will be higher and for non-reflective
surfaces (black colored), the maximum range will be lower. This can in turn
be used for detecting white/black lines (in line follower ROBOTs) or
bright/dark objects (in object identification ROBOTs)

The library provided is libupm-elementzir.so
The example provided is elementzir.cxx where it will print "Obstacle detected"
whenever an obstacle is detected by the sensor.

The image of the sensor is at docs/images/elementzir.jpg

Signed-off-by: Saloni Jain <saloni.jain@tcs.com>
Signed-off-by: Niti Rohilla <niti.rohilla@tcs.com>