Java: Added an unresolved symbol check for Java packages
Tell the linker to error on unresolved symbols.  This enables future
Java work (removing library source from Java packages) by flagging
missing reference.s

TODO: Move this up one level to all libraries.  The reason this was not
done in the same commit is that the NodeJs libraries contain unresolved
references which must come from node but remain unresolved in the NodeJs
packages.

Signed-off-by: Noel Eck <noel.eck@intel.com>